코드 클래스로 나눠 분리함.
