+++ app/src/main/java/com/example/kioskproject/DetailMenu.kt
@@ -1,0 +1,4 @@
+package com.example.kioskproject
+
+class DetailMenu {
+}
