+++ app/src/main/java/com/example/kioskproject/Menu.kt
@@ -1,101 +1,6 @@
 package com.example.kioskproject
 
-import kotlin.system.exitProcess
-
-data class MenuItem(
-    var id: Int,
-    var name: String,
-    var charge: Int
-)
-
-fun main() {
-
-    println("안녕하세요? 투썸플레이스입니다.\n")
-    Order().firstOrder()
-
-}
-
-lateinit var method: String
-lateinit var detailMenu: String
-var totalCharge: Int = 0
-
-class Order {
-
-    fun firstOrder() {
-
-        while (true) {
-            try {
-                println(
-                    "주문을 하시려면 1번을 눌러주세요.\n" +
-                            "1. 주문하기\n" +
-                            "2. 프로그램 종료."
-                )
-
-                when (readLine()?.toInt()) {
-                    1 -> {
-                        UseCafe().useMethodChoice()
-                        break
-                    }
-
-                    2 -> {
-                        println("프로그램 종료")
-                        System.exit(0)
-                    }
-
-                    else -> println("올바르지 않은 번호입니다. 다시 입력해주세요. \n----------------------------------\n")
-
-                }
-            } catch (e: java.lang.NumberFormatException) {
-                println("메뉴 주문은 숫자만 입력할 수 있습니다. \n----------------------------------\n")
-            }
-        }
-
-    }
-}
-
-class UseCafe() {
-    fun useMethodChoice() {
-
-        while (true) {
-
-            println(
-                "매장 이용 방법을 선택해주세요\n" +
-                        "1. 매장식사\n" +
-                        "2. 포장하기\n" +
-                        "3. 프로그램종료"
-            )
-
-            try {
-
-                when (readLine()?.toInt()) {
-                    1 -> {
-                        method = "매장식사"
-                        MainMenu().mainMenuList()
-                        break
-                    }
-
-                    2 -> {
-                        method = "포장하기"
-                        MainMenu().mainMenuList()
-                        break
-                    }
-
-                    3 -> {
-                        println("프로그램 종료")
-                        exitProcess(0)
-                    }
-
-                    else -> println("올바르지 않은 번호입니다. 다시 입력해주세요. \n----------------------------------\n")
-
-                }
-            } catch (e: java.lang.NumberFormatException) {
-                println("메뉴 주문은 숫자만 입력할 수 있습니다. \n----------------------------------\n")
-            }
-        }
-    }
-}
-
-class MainMenu() {
+class Menu() {
 
     fun mainMenuList() {
         while (true) {
@@ -135,7 +40,7 @@ class MainMenu() {
                         } catch (e: java.lang.NumberFormatException) {
                             println("메뉴 주문은 숫자만 입력할 수 있습니다. \n----------------------------------\n")
                         }
-                        DetailMenu().numClick()
+                        Basket().payment()
                         break
                     }
 
@@ -159,7 +64,7 @@ class MainMenu() {
                         } catch (e: java.lang.NumberFormatException) {
                             println("메뉴 주문은 숫자만 입력할 수 있습니다. \n----------------------------------\n")
                         }
-                        DetailMenu().numClick()
+                        Basket().payment()
                         break
                     }
 
@@ -182,7 +87,7 @@ class MainMenu() {
                         } catch (e: java.lang.NumberFormatException) {
                             println("메뉴 주문은 숫자만 입력할 수 있습니다. \n----------------------------------\n")
                         }
-                        DetailMenu().numClick()
+                        Basket().payment()
                         break
                     }
 
@@ -205,7 +110,7 @@ class MainMenu() {
                         } catch (e: java.lang.NumberFormatException) {
                             println("메뉴 주문은 숫자만 입력할 수 있습니다. \n----------------------------------\n")
                         }
-                        DetailMenu().numClick()
+                        Basket().payment()
                         break
                     }
 
@@ -229,7 +134,7 @@ class MainMenu() {
                         } catch (e: java.lang.NumberFormatException) {
                             println("메뉴 주문은 숫자만 입력할 수 있습니다. \n----------------------------------\n")
                         }
-                        DetailMenu().numClick()
+                        Basket().payment()
                         break
                     }
 
@@ -246,16 +151,19 @@ class MainMenu() {
                         try {
                             while (true) {
                                 val num = readLine()?.toInt()
-                                detailMenu = list[num?.minus(1)!!].name //코드 수정해보기
-                                totalCharge = list[num - 1].charge //코드 수정해보기
+                                detailMenu = list[num?.minus(1)!!].name
+                                totalCharge = list[num - 1].charge
                                 break
                             }
+                            //문자 null값 에러코드 처리방법 없음 -> 추가예정
 
+                        } catch (e: java.lang.NumberFormatException) {
+                            println("메뉴 주문은 숫자만 입력할 수 있습니다. \n----------------------------------\n")
                         } catch (e: java.lang.NumberFormatException) {
                             println("메뉴 주문은 숫자만 입력할 수 있습니다. \n----------------------------------\n")
                         }
 
-                        DetailMenu().numClick()
+                        Basket().payment()
                         break
                     }
 
@@ -267,17 +175,5 @@ class MainMenu() {
 
 
         }
-    }
-}
-
-class DetailMenu() {
-    fun numClick() {
-        println(
-            "----------------------------------\n" +
-                    "주문 내역 입니다.\n" +
-                    "$method\n" +
-                    "$detailMenu\n" +
-                    "$totalCharge"
-        )
     }
 }
