+++ app/src/main/java/com/example/kioskproject/MenuItem.kt
@@ -1,0 +1,7 @@
+package com.example.kioskproject
+
+data class MenuItem(
+    var id: Int,
+    var name: String,
+    var charge: Int
+)
