+++ app/src/main/java/com/example/kioskproject/ui/MainActivity.kt
@@ -1,11 +1,14 @@
-package com.example.kioskproject
+package com.example.kioskproject.ui
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.example.kioskproject.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
     }
+
+
 }
